% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b2p.R
\name{b2p}
\alias{b2p}
\title{Convert Bracken Output to Phyloseq Object}
\usage{
b2p(file, database, mdata = NULL)
}
\arguments{
\item{file}{Path to the Bracken output file (tab-delimited).}

\item{database}{Path to the \code{taxonomizr} database (e.g., \code{accessionTaxa.sql}). Run setupdb() first.}

\item{mdata}{Optional. Path to a metadata file (CSV with semicolon as separator) to merge into the \code{phyloseq} object. Names should be equal to bracken input files.}
}
\value{
A \code{phyloseq} object containing the OTU table and taxonomy table, and optionally, sample metadata.
}
\description{
This function converts a Bracken output file into a \code{phyloseq} object by linking the taxonomy id provided by bracken
with a NCBI database provided by\code{taxonomizr}and optionally adding metadata.
}
\examples{
\dontrun{
# Example usage
bracken <- b2p(file = "bracken.tsv", database = "database/accessionTaxa.sql")

# Example with metadata
bracken <- b2p(file = "bracken.tsv", database = "database/accessionTaxa.sql", mdata = "metadata.csv")
}
}
