% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupdb.R
\name{setupdb}
\alias{setupdb}
\title{Setup Taxonomizr Database}
\usage{
setupdb(path = "database/")
}
\arguments{
\item{path}{Directory to store the database. Default is "database/".}
}
\value{
None
}
\description{
This function sets up the \code{taxonomizr} database in a specified directory.
}
